requirements.txt
asgiref==3.8.1
Django==3.2.16
django-ckeditor==6.7.1
django-cors-headers==3.13.0
django-environ==0.11.2
django-js-asset==2.2.0
django-storages==1.14.4
djangorestframework==3.14.0
pillow==10.4.0
psycopg2==2.9.5
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
